Move step() to double for now.

+++ src/envs/env.hpp
@@ -20,7 +20,7 @@ class Env {
     int max_steps;
 
     virtual std::tuple<std::vector<float>, double, bool> step(const int &action) {return {};};
-    virtual std::tuple<std::vector<float>, double, bool> step(std::vector<float> &action) {return {};};
+    virtual std::tuple<std::vector<float>, double, bool> step(std::vector<double> &action) {return {};};
     virtual std::vector<float> reset(std::mt19937 &generator) {return {};};
 };
 #endif
